requirements.txt
tensorflow-gpu
matplotlib
pillow
opencv-python
